avoid error in wpdb::prepare with wp53
avoid "Declaration of W3TC\DbCache_Wpdb::prepare($query, $args) should
be compatible with wpdb::prepare($query, …$args)" in wp53
